enhance meta define api
